Home,Up,Past y Details crean cards desde fetch
