+++ assets/pages/details.html
@@ -39,8 +39,7 @@
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.1/dist/js/bootstrap.bundle.min.js"
         integrity="sha384-HwwvtgBNo3bZJJLYd8oVXjrBZt8cqVSpeBNS5n7C8IVInixGAoxmnlMuBnhbgrkm"
         crossorigin="anonymous"></script>
-    <script src="../scripts/data.js"></script>
-    <script src="../scripts/details.js"></script>
+    <script type="module" src="../scripts/details.js"></script>
 </body>
 
 </html>
