+++ assets/pages/upcomingEvents.html
@@ -42,8 +42,7 @@
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.1/dist/js/bootstrap.bundle.min.js"
         integrity="sha384-HwwvtgBNo3bZJJLYd8oVXjrBZt8cqVSpeBNS5n7C8IVInixGAoxmnlMuBnhbgrkm"
         crossorigin="anonymous"></script>
-    <script src="../scripts/data.js"></script>
-    <script src="../scripts/upComingEvents.js"></script>
+    <script type="module" src="../scripts/upComingEvents.js"></script>
 </body>
 
 </html>
