+++ index.html
@@ -42,8 +42,7 @@
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.1/dist/js/bootstrap.bundle.min.js"
         integrity="sha384-HwwvtgBNo3bZJJLYd8oVXjrBZt8cqVSpeBNS5n7C8IVInixGAoxmnlMuBnhbgrkm"
         crossorigin="anonymous"></script>
-    <script src="./assets/scripts/data.js"></script>
-    <script src="./assets/scripts/home.js"></script>
+    <script type="module" src="./assets/scripts/home.js"></script>
 </body>
 
 </html>
